Run pinned read regressions in native Any Quote CI suite
